Improve path detour calculation and tests
